modify condition if

+++ backend/controllers/sauce.js
@@ -49,16 +49,16 @@ exports.deleteSauce = (req, res, next) => {
         if (err) {
             return res.status(500).json({ err });
         }
-        //Impossible de vérifier l'id de l'utilisateur pour le moment (frontend)
-        /* if(req.body.userId !== sauce.userId) {
-             return res.status(403).json({ error : 'Vous n\'êtes pas autorisé à supprimer la sauce' });
-         }*/
-        const filename = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.deleteOne({ _id: req.params.id })
-                .then(() => res.status(200).json({ message: 'La sauce a été supprimée !' }))
-                .catch(error => res.status(400).json({ error }));
-        });
+        if ((userId => userId == req.body.userId)) {
+            const filename = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.deleteOne({ _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'La sauce a été supprimée !' }))
+                    .catch(error => res.status(400).json({ error }));
+            });
+        } else {
+            return res.status(403).json({ error: 'Vous n\'êtes pas autorisé à supprimer la sauce' });
+        }
     })
 };
 
@@ -67,12 +67,19 @@ exports.deleteSauce = (req, res, next) => {
 exports.likeSauce = (req, res, next) => {
     switch (req.body.like) {
         case 1:
-            Sauce.updateOne({ _id: req.params.id }, {
-                $inc: { likes: 1 }, //incrémente de 1 le nombre de likes
-                $push: { usersLiked: req.body.userId } //ajoute l'userId aux likes 
-            })
-                .then(() => res.status(200).json({ message: "L'utilisateur ajoute son like" }))
-                .catch(error => res.status(400).json({ error }));
+            Sauce.findById(req.params.id, (err, sauceResolve) => {
+                if (err) {
+                    return res.status(400).json({ error });
+                }
+                if ((sauceResolve.usersLiked.findIndex(userId => userId == req.body.userId))) {
+                    Sauce.updateOne({ _id: req.params.id }, {
+                        $inc: { likes: 1 }, //incrémente de 1 le nombre de likes
+                        $push: { usersLiked: req.body.userId } //ajoute l'userId aux likes 
+                    })
+                        .then(() => res.status(200).json({ message: "L'utilisateur ajoute son like" }))
+                        .catch(error => res.status(400).json({ error }));
+                }
+            });
             break;
 
         case 0:
@@ -99,11 +106,18 @@ exports.likeSauce = (req, res, next) => {
             break;
 
         case -1:
-            Sauce.updateOne({ _id: req.params.id }, {
-                $inc: { dislikes: 1 }, //incrémente de 1 le nombre de dislikes
-                $push: { usersDisliked: req.body.userId } //ajoute l'userId aux dislikes 
-            })
-                .then(() => res.status(200).json({ message: "L'utilisateur ajoute son dislike" }))
-                .catch(error => res.status(400).json({ error }));
+            Sauce.findById(req.params.id, (err, sauceResolve) => {
+                if (err) {
+                    return res.status(400).json({ error });
+                }
+                if ((sauceResolve.usersLiked.findIndex(userId => userId == req.body.userId))) {
+                    Sauce.updateOne({ _id: req.params.id }, {
+                        $inc: { dislikes: 1 }, //incrémente de 1 le nombre de dislikes
+                        $push: { usersDisliked: req.body.userId } //ajoute l'userId aux dislikes 
+                    })
+                        .then(() => res.status(200).json({ message: "L'utilisateur ajoute son dislike" }))
+                        .catch(error => res.status(400).json({ error }));
+                }
+            });
     }
 }
